processing-engine/requirements.txt
flask==2.0.1
werkzeug==2.0.2
requests==2.26.0
kafka-python==2.0.2
numpy==1.21.0
